Refactor template macro engine for clarity and robustness
Splits notebook-utils.ts (940 lines) into three focused modules:
notebook-utils.ts (cell traversal + sanitization), macro-substitution.ts
(legacy SQL path), template-evaluator.ts (the new walker). The original
file is now a re-export barrel so the dozen+ existing callers don't move.

Hardens useColumnWarnings to hash overridesSource by content, removing
the "must pass a stable reference" footgun the docstring previously
warned about. Adds a regression test for an unmemoized overrides array.

Extracts <WarningAwareSortHeaderRow> in table-utils.tsx so TableCell
and TableRenderer no longer each duplicate ~25 lines of SortHeader +
warning-icon wiring.
